requirements.txt
sentient-agent-framework>=0.1.0
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
fireworks-ai>=0.15.0
openai>=1.3.0
anthropic>=0.13.0
aiohttp>=3.9.1
pydantic>=2.5.0